+++ extract_msg/msg_classes/contact.py
@@ -709,8 +709,8 @@ class Contact(MessageBase):
                 'Email3 Display As': self.email3DisplayName,
             },
             '-other-': {
-                'Birthday': self.birthday.__format__('%B %d, %Y') if self.birthdayLocal else None,
-                'Anniversary': self.weddingAnniversary.__format__('%B %d, %Y') if self.weddingAnniversaryLocal else None,
+                'Birthday': self.birthday.__format__(self.dateFormat) if self.birthdayLocal else None,
+                'Anniversary': self.weddingAnniversary.__format__(self.dateFormat) if self.weddingAnniversaryLocal else None,
                 'Spouse/Partner': self.spouseName,
                 'Profession': self.profession,
                 'Children': strListToStr(self.childrensNames),