+++ extract_msg/msg_classes/journal.py
@@ -11,6 +11,7 @@ from typing import List, Optional
 from ..constants import HEADER_FORMAT_TYPE, ps
 from ..enums import LogFlags
 from .message_base import MessageBase
+from ..utils import minutesToDurationStr
 
 
 class Journal(MessageBase):
@@ -57,11 +58,11 @@ class Journal(MessageBase):
         return bool(self.getNamedProp('870F', ps.PSETID_LOG))
 
     @functools.cached_property
-    def logDuration(self) -> Optional[int]:
+    def logDuration(self) -> int:
         """
         The duration, in minutes, of the activity.
         """
-        return self.getNamedProp('8707', ps.PSETID_LOG)
+        return self.getNamedProp('8707', ps.PSETID_LOG, 0)
 
     @functools.cached_property
     def logEnd(self) -> Optional[datetime.datetime]:
@@ -100,38 +101,15 @@ class Journal(MessageBase):
 
     @property
     def headerFormatProperties(self) -> HEADER_FORMAT_TYPE:
-        """
-        Returns a dictionary of properties, in order, to be formatted into the
-        header. Keys are the names to use in the header while the values are one
-        of the following:
-        None: Signifies no data was found for the property and it should be
-            omitted from the header.
-        str: A string to be formatted into the header using the string encoding.
-        Tuple[Union[str, None], bool]: A string should be formatted into the
-            header. If the bool is True, then place an empty string if the value
-            is None, otherwise follow the same behavior as regular None.
-
-        Additional note: If the value is an empty string, it will be dropped as
-        well by default.
-
-        Additionally you can group members of a header together by placing them
-        in an embedded dictionary. Groups will be spaced out using a second
-        instance of the join string. If any member of a group is being printed,
-        it will be spaced apart from the next group/item.
-
-        If you class should not do *any* header injection, return None from this
-        property.
-        """
         return {
             '-main details-': {
-                'From': self.sender,
-                'Posted At': self.date,
-                'Conversation': self.conversation,
-            },
-            '-subject-': {
                 'Subject': self.subject,
+                'Entry Type': self.logTypeDesc,
+                'Company': self.companies[0] if self.companies else None,
             },
-            '-importance-': {
-                'Importance': self.importanceString,
-            },
+            '-time-': {
+                'Start': self.logStart.__format__(self.datetimeFormat),
+                'End': self.logEnd.__format__(self.datetimeFormat),
+                'Duration': minutesToDurationStr(self.duration),
+            }
         }