+++ extract_msg/msg_classes/meeting_cancellation.py
@@ -69,8 +69,8 @@ class MeetingCancellation(MeetingRelated):
                 'Location': self.location,
             },
             '-date-': {
-                'Start': self.startDate.__format__('%a, %d %b %Y %H:%M %z') if self.startDate else None,
-                'End': self.endDate.__format__('%a, %d %b %Y %H:%M %z') if self.endDate else None,
+                'Start': self.startDate.__format__(self.datetimeFormat) if self.startDate else None,
+                'End': self.endDate.__format__(self.datetimeFormat) if self.endDate else None,
                 'Show Time As': 'Free',
             },
             '-recurrence-': {